requirements.txt
Flask==1.1.2
Flask_Admin==1.5.7
Flask_SQLAlchemy==2.4.4
PyJWT==2.1.0
gunicorn
